Add support for EMM reassemblation.
OSCAM does the reassemblation only when the input is dvbapi.
This means that a client (like tsdecrypt) that uses network
protocol must do EMM reassembly by itself. This commit adds
reassembly support for Cryptoworks and Viaccess CA systems.

The card update was tested and found to be working on Swiss
Viaccess cards.